Move settings-datastore test helpers to dedicated testing module
    - Relocate `FakeDataStoreModule` and `FakeJcaSettingsSerializer` to
      the `:data:settings-datastore:testing` module.
    - Update package declarations, imports, and build dependencies
  accordingly.
