+++ data/settings-datastore/testing/src/main/java/com/google/jetpackcamera/settingsdatastore/testing/FakeJcaSettingsSerializer.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.jetpackcamera.data.settingsdatastore.test
+package com.google.jetpackcamera.settingsdatastore.testing
 
 import androidx.datastore.core.CorruptionException
 import androidx.datastore.core.Serializer
